+++ mycodo/mycodo_flask/utils/utils_function.py
@@ -11,15 +11,17 @@ from mycodo.databases.models import Actions
 from mycodo.databases.models import Camera
 from mycodo.databases.models import Conditional
 from mycodo.databases.models import DisplayOrder
+from mycodo.databases.models import Function
 from mycodo.databases.models import PID
 from mycodo.databases.models import Trigger
 from mycodo.mycodo_client import DaemonControl
 from mycodo.mycodo_flask.extensions import db
 from mycodo.mycodo_flask.utils.utils_general import add_display_order
 from mycodo.mycodo_flask.utils.utils_general import delete_entry_with_id
-from mycodo.mycodo_flask.utils.utils_general import flash_form_errors
 from mycodo.mycodo_flask.utils.utils_general import flash_success_errors
+from mycodo.mycodo_flask.utils.utils_general import reorder
 from mycodo.utils.system_pi import csv_to_list_of_str
+from mycodo.utils.system_pi import list_to_csv
 from mycodo.utils.system_pi import str_is_float
 
 logger = logging.getLogger(__name__)
@@ -29,52 +31,92 @@ logger = logging.getLogger(__name__)
 # Function manipulation
 #
 
-def func_add(form_add_func):
+def function_add(form_add_func):
     action = '{action} {controller}'.format(
         action=gettext("Add"),
         controller=gettext("Function"))
     error = []
 
-    if form_add_func.validate():
-        try:
-            if form_add_func.func_type.data == 'pid':
-                new_func = PID().save()
-                if not error:
-                    display_order = csv_to_list_of_str(
-                        DisplayOrder.query.first().pid)
-                    DisplayOrder.query.first().pid = add_display_order(
-                        display_order, new_func.unique_id)
-                    db.session.commit()
-            elif form_add_func.func_type.data == 'conditional':
-                new_func = Conditional()
-                new_func.save()
-                if not error:
-                    display_order = csv_to_list_of_str(
-                        DisplayOrder.query.first().conditional)
-                    DisplayOrder.query.first().conditional = add_display_order(
-                        display_order, new_func.unique_id)
-                    db.session.commit()
-            else:
-                new_func = Trigger()
-                for id, name in FUNCTION_TYPES:
-                    if form_add_func.func_type.data == id:
-                        new_func.name = name
-                new_func.trigger_type = form_add_func.func_type.data
-                new_func.save()
-                if not error:
-                    display_order = csv_to_list_of_str(
-                        DisplayOrder.query.first().trigger)
-                    DisplayOrder.query.first().trigger = add_display_order(
-                        display_order, new_func.unique_id)
-                    db.session.commit()
+    new_func = None
 
-        except sqlalchemy.exc.OperationalError as except_msg:
-            error.append(except_msg)
-        except sqlalchemy.exc.IntegrityError as except_msg:
-            error.append(except_msg)
-        flash_success_errors(error, action, url_for('routes_page.page_function'))
-    else:
-        flash_form_errors(form_add_func)
+    try:
+        if form_add_func.func_type.data.startswith('conditional_'):
+            new_func = Conditional().save()
+        elif form_add_func.func_type.data.startswith('pid_'):
+            new_func = PID().save()
+        elif form_add_func.func_type.data.startswith('trigger_'):
+            new_func = Trigger().save()
+            for id, name, _ in FUNCTION_TYPES:
+                if form_add_func.func_type.data == id:
+                    new_func.name = name
+            new_func.trigger_type = form_add_func.func_type.data
+            new_func.save()
+        elif form_add_func.func_type.data.startswith('function_'):
+            new_func = Function().save()
+        else:
+            error.append("Unknown function type: '{}'".format(
+                form_add_func.func_type.data))
+
+        if not error:
+            display_order = csv_to_list_of_str(
+                DisplayOrder.query.first().function)
+            DisplayOrder.query.first().function = add_display_order(
+                display_order, new_func.unique_id)
+            db.session.commit()
+
+    except sqlalchemy.exc.OperationalError as except_msg:
+        error.append(except_msg)
+    except sqlalchemy.exc.IntegrityError as except_msg:
+        error.append(except_msg)
+    except Exception as except_msg:
+        error.append(except_msg)
+
+    flash_success_errors(error, action, url_for('routes_page.page_function'))
+
+
+def function_mod(form):
+    """Modify a Function"""
+    error = []
+    action = '{action} {controller}'.format(
+        action=gettext("Mod"),
+        controller=(gettext("Function")))
+
+    try:
+        function_mod = Function.query.filter(
+            Function.unique_id == form.function_id.data).first()
+
+        function_mod.name = form.name.data
+
+        if not error:
+            db.session.commit()
+
+    except sqlalchemy.exc.OperationalError as except_msg:
+        error.append(except_msg)
+    except sqlalchemy.exc.IntegrityError as except_msg:
+        error.append(except_msg)
+    except Exception as except_msg:
+        error.append(except_msg)
+
+    flash_success_errors(error, action, url_for('routes_page.page_function'))
+
+
+def function_del(function_id):
+    action = '{action} {controller}'.format(
+        action=gettext("Delete"),
+        controller=gettext("Function"))
+    error = []
+
+    try:
+        delete_entry_with_id(Function, function_id)
+
+        display_order = csv_to_list_of_str(DisplayOrder.query.first().function)
+        display_order.remove(function_id)
+        DisplayOrder.query.first().function = list_to_csv(display_order)
+        db.session.commit()
+    except Exception as except_msg:
+        error.append(except_msg)
+
+    flash_success_errors(error, action, url_for('routes_page.page_function'))
 
 
 def action_add(form):
@@ -116,43 +158,6 @@ def action_add(form):
     flash_success_errors(error, action, url_for('routes_page.page_function'))
 
 
-def action_test_all(form):
-    """Test All Conditional Actions"""
-    error = []
-
-    func_type = None
-    func = None
-
-    if form.function_type.data == 'conditional':
-        func_type = gettext("Conditional")
-        func = Conditional.query.filter(
-            Conditional.unique_id == form.function_id.data).first()
-    elif form.function_type.data == 'trigger':
-        func_type = gettext("Trigger")
-        func = Trigger.query.filter(
-            Trigger.unique_id == form.function_id.data).first()
-    else:
-        error.append("Unknown Function type")
-
-    action = '{action} {controller}'.format(
-        action=gettext("Test All"),
-        controller='{} {}'.format(func_type, gettext("Action")))
-
-    if func and not func.is_activated:
-        error.append("Activate the Conditional before testing all Actions")
-
-    try:
-        if not error:
-            control = DaemonControl()
-            control.test_trigger_actions(
-                form.function_id.data,
-                message="Test triggering all actions of function {}".format(form.function_id.data))
-    except Exception as except_msg:
-        error.append(except_msg)
-
-    flash_success_errors(error, action, url_for('routes_page.page_function'))
-
-
 def action_mod(form):
     """Modify a Conditional Action"""
     error = []
@@ -166,12 +171,15 @@ def action_mod(form):
         mod_action = Actions.query.filter(
             Actions.unique_id == form.function_action_id.data).first()
 
-        if mod_action.action_type == 'output':
+        if mod_action.action_type == 'pause_actions':
+            mod_action.pause_duration = form.pause_duration.data
+
+        elif mod_action.action_type == 'output':
             mod_action.do_unique_id = form.do_unique_id.data
             mod_action.do_output_state = form.do_output_state.data
             mod_action.do_output_duration = form.do_output_duration.data
 
-        if mod_action.action_type == 'output_pwm':
+        elif mod_action.action_type == 'output_pwm':
             mod_action.do_unique_id = form.do_unique_id.data
             mod_action.do_output_pwm = form.do_output_pwm.data
 
@@ -264,6 +272,62 @@ def action_del(form):
     flash_success_errors(error, action, url_for('routes_page.page_function'))
 
 
+def action_test_all(form):
+    """Test All Conditional Actions"""
+    error = []
+
+    func_type = None
+    func = None
+
+    if form.function_type.data == 'conditional':
+        func_type = gettext("Conditional")
+        func = Conditional.query.filter(
+            Conditional.unique_id == form.function_id.data).first()
+    elif form.function_type.data == 'trigger':
+        func_type = gettext("Trigger")
+        func = Trigger.query.filter(
+            Trigger.unique_id == form.function_id.data).first()
+    else:
+        error.append("Unknown Function type")
+
+    action = '{action} {controller}'.format(
+        action=gettext("Test All"),
+        controller='{} {}'.format(func_type, gettext("Action")))
+
+    if func and not func.is_activated:
+        error.append("Activate the Conditional before testing all Actions")
+
+    try:
+        if not error:
+            control = DaemonControl()
+            control.test_trigger_actions(
+                form.function_id.data,
+                message="Test triggering all actions of function {}".format(form.function_id.data))
+    except Exception as except_msg:
+        error.append(except_msg)
+
+    flash_success_errors(error, action, url_for('routes_page.page_function'))
+
+
+def function_reorder(function_id, display_order, direction):
+    action = '{action} {controller}'.format(
+        action=gettext("Reorder"),
+        controller=gettext("Function"))
+    error = []
+    try:
+        status, reord_list = reorder(display_order,
+                                     function_id,
+                                     direction)
+        if status == 'success':
+            DisplayOrder.query.first().function = ','.join(map(str, reord_list))
+            db.session.commit()
+        else:
+            error.append(reord_list)
+    except Exception as except_msg:
+        error.append(except_msg)
+    flash_success_errors(error, action, url_for('routes_page.page_function'))
+
+
 def check_form_actions(form, error):
     """Check if the Actions form inputs are valid"""
     action = Actions.query.filter(
@@ -319,7 +383,7 @@ def check_actions(action, error):
         if not action.do_output_state or action.do_output_state == '':
             error.append("State must be set")
     elif action.action_type in ['activate_pid',
-                                   'deactivate_pid']:
+                                'deactivate_pid']:
         if not action.do_unique_id or action.do_unique_id == '':
             error.append("PID must be set")
     elif action.action_type == 'email':